+++ R/InsuranceTarif.R
@@ -236,6 +236,12 @@ InsuranceTarif = R6Class(
 
       # Fill all remaining uninitialized values with their defaults, except for profit participation params
       self$Parameters = InsuranceContract.ParametersFallback(self$Parameters, InsuranceContract.ParameterDefaults, ppParameters = FALSE);
+
+      # Properly set up the rounding helper
+      if (is.null(self$Parameters$Hooks$Rounding)) self$Parameters$Hooks$Rounding = list()
+      if (is.list(self$Parameters$Hooks$Rounding)){
+        self$Parameters$Hooks$Rounding = do.call(RoundingHelper$new, self$Parameters$Hooks$Rounding)
+      }
     },
 
     #' @description create a copy of a tariff with certain parameters changed
@@ -271,6 +277,11 @@ InsuranceTarif = R6Class(
       if (!missing(tariffType)) cloned$tariffType = tariffType;
 
       cloned$Parameters = InsuranceContract.ParametersFill(cloned$Parameters, ...);
+      # Properly set up the rounding helper
+      if (is.null(cloned$Parameters$Hooks$Rounding)) cloned$Parameters$Hooks$Rounding = list()
+      if (is.list(cloned$Parameters$Hooks$Rounding)){
+        cloned$Parameters$Hooks$Rounding = do.call(RoundingHelper$new, cloned$Parameters$Hooks$Rounding)
+      }
       cloned
     },
 
@@ -364,7 +375,8 @@ InsuranceTarif = R6Class(
         } else {
           px = 1 - qx
         }
-        df = data.frame(age = ages, qx = qx, ix = ix, px = px, row.names = ages - age)
+        rd = params$Hooks$Rounding
+        df = data.frame(age = ages, qx = rd$round("qx", qx), ix = rd$round("ix", ix), px = rd$round("px", px), row.names = ages - age)
         df
       }
     },
@@ -727,6 +739,7 @@ InsuranceTarif = R6Class(
       if (getOption('LIC.debug.presentValueCashFlows', FALSE)) {
         browser();
       }
+      rd = params$Hooks$Rounding
 
       qq = self$getTransitionProbabilities(params, values);
 
@@ -746,17 +759,17 @@ InsuranceTarif = R6Class(
         (values$cashFlows[,"death_GrossPremium"] - values$cashFlows[,"premiums_advance"]);
 
       pv = cbind(
-        premiums = pvf$survival(values$cashFlows$premiums_advance, values$cashFlows$premiums_arrears,
-          m = params$ContractData$premiumFrequency, mCorrection = premiumFreqCorr),
-        additional_capital = pvf$survival(advance = values$cashFlows$additional_capital),
-        guaranteed = pvf$guaranteed(values$cashFlows$guaranteed_advance, values$cashFlows$guaranteed_arrears),
-        survival = pvf$survival(values$cashFlows$survival_advance, values$cashFlows$survival_arrears),
-        death_SumInsured = pvf$death(values$cashFlows$death_SumInsured),
-        disease_SumInsured = pvf$disease(values$cashFlows$disease_SumInsured),
-        death_GrossPremium = pvRefund,
-        death_Refund_past = pvRefundPast,
-        death_Refund_future = pvRefund - pvRefundPast,
-        death_PremiumFree = pvf$death(values$cashFlows$death_PremiumFree)
+        premiums = rd$round("PV Premiums", pvf$survival(values$cashFlows$premiums_advance, values$cashFlows$premiums_arrears,
+          m = params$ContractData$premiumFrequency, mCorrection = premiumFreqCorr)),
+        additional_capital = rd$round("PV AdditionalCapital", pvf$survival(advance = values$cashFlows$additional_capital)),
+        guaranteed = rd$round("PV Guarantee", pvf$guaranteed(values$cashFlows$guaranteed_advance, values$cashFlows$guaranteed_arrears)),
+        survival = rd$round("PV Survival", pvf$survival(values$cashFlows$survival_advance, values$cashFlows$survival_arrears)),
+        death_SumInsured = rd$round("PV Death", pvf$death(values$cashFlows$death_SumInsured)),
+        disease_SumInsured = rd$round("PV Disease", pvf$disease(values$cashFlows$disease_SumInsured)),
+        death_GrossPremium = rd$round("PV Death PremiumRefund", pvRefund),
+        death_Refund_past = rd$round("PV Death PremiumRefund Past", pvRefundPast),
+        death_Refund_future = rd$round("PV Death PremiumRefund Future", pvRefund - pvRefundPast),
+        death_PremiumFree = rd$round("PV Death PremiumFree", pvf$death(values$cashFlows$death_PremiumFree))
       );
 
       rownames(pv) <- pad0(rownames(qq), values$int$l);
@@ -817,6 +830,8 @@ InsuranceTarif = R6Class(
             pvfben$disease(cf$disease_SumInsured * cfCosts);
       }
 
+      rd = params$Hooks$Rounding;
+      pvc = rd$round("PV Costs", pvc)
       applyHook(hook = params$Hooks$adjustPresentValuesCosts, val = pvc, params = params, values = values, presentValues = presentValues)
     },
 
@@ -831,6 +846,7 @@ InsuranceTarif = R6Class(
       if (getOption('LIC.debug.getAbsCashFlows', FALSE)) {
         browser();
       }
+      rd = params$Hooks$Rounding;
 
       # TODO: Set up a nice list with coefficients for each type of cashflow,
       # rather than multiplying each item manually (this also mitigates the risk
@@ -854,6 +870,8 @@ InsuranceTarif = R6Class(
       values$cashFlows[,"death_SumInsured"] = values$cashFlows[,"death_SumInsured"] + values$cashFlows[,"death_GrossPremium"]
       colnames(values$cashFlows)[colnames(values$cashFlows) == "death_SumInsured"] = "death";
       # cashFlows[,"death_GrossPremium"] = NULL;
+      values$cashFlows = rd$round("CF absolute", values$cashFlows);
+
 
       # costs relative to sumInsured are already set up as the correct multiple
       # of the original SI, including the dynamic changes over time!
@@ -863,6 +881,7 @@ InsuranceTarif = R6Class(
         values$cashFlowsCosts[,,"NetPremium",] * values$premiums[["net"]] +
         # values$cashFlowsCosts[,,"Benefits",] * TODO!!!
         values$cashFlowsCosts[,,"Constant",];
+      values$cashFlowsCosts = rd$round("CF costs absolute", values$cashFlowsCosts);
 
       # Handle survival CF differently, because we don't want ".survival" in the column names!
       cbind(values$cashFlows, values$cashFlowsCosts[,,"survival"], values$cashFlowsCosts[,,-1])
@@ -897,6 +916,8 @@ InsuranceTarif = R6Class(
       pv[,"death_SumInsured"] = pv[,"death_SumInsured"] + pv[,"death_GrossPremium"]
       colnames(pv)[colnames(pv) == "death_SumInsured"] = "death";
 
+      pv = params$Hooks$Rounding$round("PV absolute", pv);
+
       cbind("premiums.unit" = values$presentValues[,"premiums"], pv)
     },
 
@@ -929,6 +950,10 @@ InsuranceTarif = R6Class(
         values$presentValuesCosts[,,"Constant",] / params$ContractData$sumInsured,
         dims = 2)
 
+      rd = params$Hooks$Rounding;
+      benefits = rd$round("PV abs benefits", benefits)
+      allBenefits = rd$round("PV abs allBenefits", allBenefits)
+      benefitsCosts = rd$round("PV abs benefitsCosts", benefitsCosts)
 
       cbind(
         benefits = benefits,
@@ -1146,7 +1171,7 @@ InsuranceTarif = R6Class(
         temp = temp /
           (enumerator / denominator * (1 + noMedicalExam.relative + extraChargeGrossPremium) + noMedicalExam - sumRebate - extraRebate);
       }
-      sumInsured = temp
+      sumInsured = params$Hooks$Rounding$round("sumInsured", temp);
 
       sumInsured
     },
@@ -1160,6 +1185,8 @@ InsuranceTarif = R6Class(
       if (getOption('LIC.debug.premiumCalculation', FALSE)) {
         browser();
       }
+      rd = params$Hooks$Rounding;
+
       loadings = params$Loadings;
       sumInsured = params$ContractData$sumInsured
       values$premiums = c(
@@ -1191,8 +1218,8 @@ InsuranceTarif = R6Class(
       coeff = self$getPremiumCoefficients("gross", pv * 0, pvCost * 0, premiums = values$premiums, params = params, values = values, premiumCalculationTime = premiumCalculationTime)
       enumerator  = sum(coeff[["SumInsured"]][["benefits"]] * pv) + sum(coeff[["SumInsured"]][["costs"]] * pvCost);
       denominator = sum(coeff[["Premium"   ]][["benefits"]] * pv) + sum(coeff[["Premium"   ]][["costs"]] * pvCost);
-      values$premiums[["unit.gross"]] = enumerator/ifelse(denominator == 0, 1, denominator) * (1 + loadings$ongoingAlphaGrossPremium);
-      values$premiums[["gross"]] = values$premiums[["unit.gross"]] * sumInsured;
+      values$premiums[["unit.gross"]] = rd$round("Premium gross unit", enumerator/ifelse(denominator == 0, 1, denominator) * (1 + loadings$ongoingAlphaGrossPremium));
+      values$premiums[["gross"]] = rd$round("Premium gross", values$premiums[["unit.gross"]] * sumInsured);
       coefficients[["gross"]] = coeff;
 
       # ======================================================================= =
@@ -1201,8 +1228,8 @@ InsuranceTarif = R6Class(
       coeff = self$getPremiumCoefficients("net", pv*0, pvCost*0, premiums = values$premiums, params = params, values = values, premiumCalculationTime = premiumCalculationTime)
       enumerator  = sum(coeff[["SumInsured"]][["benefits"]] * pv) + sum(coeff[["SumInsured"]][["costs"]] * pvCost);
       denominator = sum(coeff[["Premium"   ]][["benefits"]] * pv) + sum(coeff[["Premium"   ]][["costs"]] * pvCost);
-      values$premiums[["unit.net"]] = enumerator/ifelse(denominator == 0, 1, denominator);
-      values$premiums[["net"]] = values$premiums[["unit.net"]] * sumInsured;
+      values$premiums[["unit.net"]] = rd$round("Premium net unit", enumerator/ifelse(denominator == 0, 1, denominator));
+      values$premiums[["net"]] = rd$round("Premium net", values$premiums[["unit.net"]] * sumInsured);
       coefficients[["net"]] = coeff;
 
       # ======================================================================= =
@@ -1211,8 +1238,8 @@ InsuranceTarif = R6Class(
       coeff = self$getPremiumCoefficients("Zillmer", pv * 0, pvCost * 0, premiums = values$premiums, params = params, values = values, premiumCalculationTime = premiumCalculationTime);
       enumerator  = sum(coeff[["SumInsured"]][["benefits"]] * pv) + sum(coeff[["SumInsured"]][["costs"]] * pvCost);
       denominator = sum(coeff[["Premium"   ]][["benefits"]] * pv) + sum(coeff[["Premium"   ]][["costs"]] * pvCost);
-      values$premiums[["unit.Zillmer"]] = enumerator/ifelse(denominator == 0, 1, denominator);
-      values$premiums[["Zillmer"]] = values$premiums[["unit.Zillmer"]] * sumInsured;
+      values$premiums[["unit.Zillmer"]] = rd$round("Premium Zillmer unit", enumerator/ifelse(denominator == 0, 1, denominator));
+      values$premiums[["Zillmer"]] = rd$round("Premium Zillmer", values$premiums[["unit.Zillmer"]] * sumInsured);
       coefficients[["Zillmer"]] = coeff;
 
 
@@ -1220,6 +1247,7 @@ InsuranceTarif = R6Class(
       # Additional premium components (after gross premium)
       # ----------------------------------------------------------------------- -
       # The written premium is the gross premium with additional loadings, rebates, unit costs and taxes
+      # TODO: Think through, how each of the components should / could be rounded
       tax           = valueOrFunction(loadings$tax,          params = params, values = values);
       unitCosts     = valueOrFunction(loadings$unitcosts,    params = params, values = values);
       noMedicalExam = valueOrFunction(loadings$noMedicalExam,params = params, values = values);
@@ -1260,10 +1288,10 @@ InsuranceTarif = R6Class(
       premiumBeforeTax = premiumBeforeTax * (1 - premiumRebate - advanceProfitParticipationUnitCosts - partnerRebate);
       premiumBeforeTax.y = premiumBeforeTax * (1 + frequencyLoading);
       premiumBeforeTax = premiumBeforeTax.y / params$ContractData$premiumFrequency;
-      values$premiums[["written_yearly"]] = premiumBeforeTax.y * (1 + tax)
-      values$premiums[["written_beforetax"]] = premiumBeforeTax;
-      values$premiums[["tax"]] = premiumBeforeTax * tax;
-      values$premiums[["written"]] = premiumBeforeTax * (1 + tax);
+      values$premiums[["written_yearly"]] = rd$round("Premium written yearly", premiumBeforeTax.y * (1 + tax))
+      values$premiums[["written_beforetax"]] = rd$round("Premium written beforeTax", premiumBeforeTax);
+      values$premiums[["tax"]] = rd$round("Premium tax", premiumBeforeTax * tax);
+      values$premiums[["written"]] = rd$round("Premium written", values$premiums[["written_beforetax"]] + values$premiums[["tax"]]);
 
       applyHook(
         params$Hooks$adjustPremiums,
@@ -1281,6 +1309,7 @@ InsuranceTarif = R6Class(
       if (getOption('LIC.debug.reserveCalculation', FALSE)) {
         browser();
       }
+      rd = params$Hooks$Rounding;
       t = "0"
       securityFactor = (1 + valueOrFunction(params$Loadings$security, params = params, values = values));
       ppScheme      = params$ProfitParticipation$profitParticipationScheme;
@@ -1288,10 +1317,10 @@ InsuranceTarif = R6Class(
       absPV = applyHook(params$Hooks$adjustPVForReserves, values$absPresentValues, params = params, values = values);
 
       # Net, Zillmer and Gross reserves
-      resNet = absPV[,"benefitsAndRefund"] * securityFactor - values$premiums[["net"]] * absPV[,"premiums.unit"];
+      resNet = rd$round("Res net", absPV[,"benefitsAndRefund"] * securityFactor - values$premiums[["net"]] * absPV[,"premiums.unit"]);
       BWZcorr = ifelse(absPV[t, "premiums"] == 0, 0,
                        absPV[t, "Zillmer"] / absPV[t, "premiums"]) * absPV[,"premiums"];
-      resZ = resNet - BWZcorr;
+      resZ = rd$round("Res Zillmer", resNet - BWZcorr);
 
       resAdeq = absPV[,"benefitsAndRefund"] * securityFactor +
           absPV[,"alpha"] + absPV[,"beta"] + absPV[,"gamma"] -
@@ -1299,10 +1328,11 @@ InsuranceTarif = R6Class(
       if (params$Features$unitcostsInGross) {
           resAdeq = resAdeq + absPV[, "unitcosts"]
       }
+      resAdeq = rd$round("Res adequate", resAdeq)
 
-      resGamma = absPV[,"gamma"] -
+      resGamma = rd$round("Res gamma", absPV[,"gamma"] -
         ifelse(absPV[t, "premiums"] == 0, 0,
-               absPV[t, "gamma"] / absPV[t, "premiums"]) * absPV[,"premiums"]
+               absPV[t, "gamma"] / absPV[t, "premiums"]) * absPV[,"premiums"]);
 
 
       advanceProfitParticipation = 0;
@@ -1339,7 +1369,10 @@ InsuranceTarif = R6Class(
         resContractual = resAdeq + resGamma
         resReduction = resAdeq + alphaRefund;
       }
-      resConversion = resContractual * (1 - advanceProfitParticipation);
+      resContractual = rd$round("Res contractual", resContractual)
+      resReduction = rd$round("Res reduction", resReduction)
+
+      resConversion = rd$round("Res conversion", resContractual * (1 - advanceProfitParticipation));
       if (params$Features$surrenderIncludesCostsReserves) {
         resReduction = resReduction + resGamma;
       }
@@ -1379,6 +1412,7 @@ InsuranceTarif = R6Class(
           partnerRebate = valueOrFunction(params$Loadings$partnerRebate, params = params, values = values);
           surrenderValue = resReduction * (1 - advanceProfitParticipationUnitCosts - partnerRebate);
       }
+      surrenderValue = rd$round("Surrender Value", surrenderValue)
 
 
       # Calculate new sum insured after premium waiver
@@ -1392,6 +1426,7 @@ InsuranceTarif = R6Class(
       newSI = ifelse(premiumfreePV == 0, 0,
         (premiumfreeValue - absPV[,"death_Refund_past"] * securityFactor - c(Storno)) /
         premiumfreePV * params$ContractData$sumInsured);
+      newSI = rd$round("Premiumfree SI", newSI)
 
       cbind(res,
             "PremiumsPaid" = Reduce("+", values$absCashFlows$premiums_advance, accumulate = TRUE),
@@ -1443,6 +1478,7 @@ InsuranceTarif = R6Class(
       if (getOption('LIC.debug.reserveCalculationBalanceSheet', FALSE)) {
         browser();
       }
+      rd = params$Hooks$Rounding
       reserves = values$reserves;
       years = length(reserves[,"Zillmer"]);
       # Balance sheet reserves:
@@ -1496,7 +1532,7 @@ InsuranceTarif = R6Class(
                   "unearned Premiums"     = unearnedPremiums
       );
       rownames(res) <- rownames(reserves);
-      res
+      rd$round("Balance Sheet", res)
     },
 
     #' @description Calculate the profit participation given the contract
@@ -1549,6 +1585,7 @@ InsuranceTarif = R6Class(
     #' @details Not to be called directly, but implicitly by the [InsuranceContract] object.
     #'          All premiums, reserves and present values have already been calculated.
     premiumDecomposition = function(params, values) {
+      # TODO: No rounding applied yet!
       if (getOption('LIC.debug.premiumDecomposition', FALSE)) {
         browser();
       }