+++ NEWS.md
@@ -2,8 +2,9 @@
 # Version 1.0.1: XXXXXXXXXXX XX, 2023
   * New parameters:
     - survivalBenefit: Generalize survival benefit vectors (previously: unit CF 1 at end of contract)
-    - gammaInZillmer: As a feature, include gamma costs (but not beta) in the Zillmer premium
-  * Improve test case generation: Als generate code to export sample contract to Excel
+    - gammaInZillmer: As a company-specific feature, include gamma costs (but not beta) in the Zillmer premium
+  * Improve test case generation: Also generate code to export sample contract to Excel
+  * Add feature to round intermediate value using a RoundingHelper class (stored in Parameters$Hooks$Rounding)
   
 
 # Version 1.0.0: October 27, 2023